requirements.txt
# ─── Core ───────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.6.4
loguru==0.7.2

# ─── HTTP / API Ingestion ────────────────────────────────────────────────────
requests==2.31.0
tenacity==8.2.3          # retry with back-off for API rate limits

# ─── Data Processing ─────────────────────────────────────────────────────────
pandas==2.2.1
numpy==1.26.4
pyarrow==15.0.2          # columnar serialisation for data lake files

# ─── Database ────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
sqlalchemy==2.0.29

# ─── Kafka ───────────────────────────────────────────────────────────────────
confluent-kafka==2.3.0

# ─── Spark (PySpark) ─────────────────────────────────────────────────────────
pyspark==3.5.1

# ─── Airflow (install separately with extras) ────────────────────────────────
# apache-airflow[postgres,http]==2.8.1

# ─── Visualisation / Architecture Diagram ────────────────────────────────────
matplotlib==3.8.4

# ─── Testing ─────────────────────────────────────────────────────────────────
pytest==8.1.1
pytest-cov==5.0.0
freezegun==1.4.0
